UPdated Directory Structure
